feat: Add login and signup pages
