--- frontend/src/components/loginPage.tsx
@@ -1,101 +1,0 @@
-import { useState } from 'react'
-import Link from 'next/link'
-
-export default function LoginPage({ onSwitch }: { onSwitch?: () => void }) {
-  const [showPassword, setShowPassword] = useState(false)
-
-  return (
-    <div className="min-h-screen flex items-center justify-center color-bg py-12 px-4 sm:px-6 lg:px-8">
-      <div className="w-full max-w-md">
-        <div className="bg-white rounded-2xl p-6 sm:p-8 flex flex-col shadow-lg">
-          <h2 className="text-center text-2xl sm:text-3xl font-semibold mb-6">Log in</h2>
-          <form action="" className="flex  flex-col gap-4">
-            <div className="flex flex-col gap-1">
-              <label htmlFor="">Email</label>
-              <input
-                type="text"
-                placeholder="Email"
-                className="border border-gray-300 rounded-lg px-4 py-2 focus:outline-none focus:ring-2 focus:ring-primary"
-              />
-            </div>
-            <div className="flex flex-col gap-1">
-              <label htmlFor="">Password</label>
-              <div className="relative">
-                <input
-                  type={showPassword ? 'text' : 'password'}
-                  placeholder="Password"
-                  className="border border-gray-300 rounded-lg px-4 py-2 pr-12 w-full focus:outline-none focus:ring-2 focus:ring-primary"
-                />
-                <button
-                  type="button"
-                  onClick={() => setShowPassword(!showPassword)}
-                  className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 transition"
-                  aria-label={showPassword ? 'Hide Password' : 'Show Password'}
-                >
-                  {showPassword ? (
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      strokeWidth={1.5}
-                      stroke="currentColor"
-                      className="w-5 h-5"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        d="M3.98 8.223A10.477 10.477 0 001.934 12C3.226 16.338 7.244 19.5 12 19.5c.993 0 1.953-.138 2.863-.395M6.228 6.228A10.45 10.45 0 0112 4.5c4.756 0 8.773 3.162 10.065 7.498a10.523 10.523 0 01-4.293 5.774M6.228 6.228L3 3m3.228 3.228l3.65 3.65m7.894 7.894L21 21m-3.228-3.228l-3.65-3.65m0 0a3 3 0 10-4.243-4.243m4.242 4.242L9.88 9.88"
-                      />
-                    </svg>
-                  ) : (
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      strokeWidth={1.5}
-                      stroke="currentColor"
-                      className="w-5 h-5"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        d="M2.036 12.322a1.012 1.012 0 010-.639C3.423 7.51 7.36 4.5 12 4.5c4.638 0 8.573 3.007 9.963 7.178.07.207.07.431 0 .639C20.577 16.49 16.64 19.5 12 19.5c-4.638 0-8.573-3.007-9.963-7.178z"
-                      />
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"
-                      />
-                    </svg>
-                  )}
-                </button>
-              </div>
-            </div>
-            {/* buttons */}
-            <div className="flex flex-col sm:flex-row gap-3 sm:gap-4 mt-2">
-              <button className="mt-2 color-primary text-white w-full sm:w-50 py-2 cursor-pointer rounded-full font-medium hover:opacity-90 transition">
-                Sign In
-              </button>
-              {onSwitch ? (
-                <button
-                  type="button"
-                  onClick={onSwitch}
-                  className="mt-2 bg-gray-200 text-gray-800 border border-gray-400 cursor-pointer py-2 w-full sm:w-50 rounded-full font-medium hover:bg-gray-300 transition"
-                >
-                  No account? Sign Up
-                </button>
-              ) : (
-                <Link
-                  href="/signup"
-                  className="mt-2 bg-gray-200 text-gray-800 border border-gray-400 cursor-pointer py-2 w-full sm:w-50 rounded-full font-medium hover:bg-gray-300 transition text-center"
-                >
-                  No account? Sign Up
-                </Link>
-              )}
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
-  )
-}
